When the kernel module is loaded with a debug mask parameter, the code accepts the parameter into a 32-bit variable, which is then cast into a 64-bit parameter.  The insmod command does not handle 64-bit input parameters, but the kernel gossip_debug_mask should be a 64-bit value to acccomodate the existing gossip-debug functions.  Modified Files:  	src/common/gossip/gossip.h src/io/dev/pint-dev.c  	src/kernel/linux-2.6/devpvfs2-req.c  	src/kernel/linux-2.6/pvfs2-mod.c
SVN revision 8229

+++ src/common/gossip/gossip.h
@@ -49,7 +49,7 @@ enum gossip_logstamp
 #ifdef GOSSIP_DISABLE_DEBUG
 #define gossip_debug(mask, format, f...) do {} while(0)
 #else
-extern int gossip_debug_mask;
+extern uint64_t gossip_debug_mask;
 
 /* try to avoid function call overhead by checking masks in macro */
 #define gossip_debug(mask, format, f...)                  \
